tools/pyannote-community-worker/requirements.txt
pyannote.audio>=4.0,<5
torch
